+++ Assets/External Assets/Entities/Npcs/medieval/Fighter/FighterDialogStartScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e17a2422d847d334db84cc00348644ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
