+++ Assets/External Assets/Entities/Npcs/medieval/Fighter/FighterDialogTown.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6626ff0d79cdc3c40ad0207e33313057
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
