+++ Assets/External Assets/Entities/Npcs/medieval/Thief/Theif.prefab
@@ -248,7 +248,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &9063631423490453358
 RectTransform:
   m_ObjectHideFlags: 0
