+++ Assets/External Assets/Entities/Npcs/medieval/Thief/ThiefDialogStartScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7250d6de5fdd91240bc3a6cd1d0d329c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
