+++ Assets/External Assets/Entities/Npcs/medieval/Thief/ThiefDialogTown.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 65640034668a0354288c9f9303f5cf5d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
