+++ Assets/External Assets/Entities/Npcs/medieval/beggar/Beggar.prefab
@@ -17,7 +17,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &7026803298345901836
 Transform:
   m_ObjectHideFlags: 0
@@ -541,6 +541,10 @@ PrefabInstance:
     - target: {fileID: 4127760927357580505, guid: 7fb4574d62b4c9e4cb95921ac1978c4f, type: 3}
       propertyPath: m_Name
       value: GUI
+      objectReference: {fileID: 0}
+    - target: {fileID: 4127760927357580505, guid: 7fb4574d62b4c9e4cb95921ac1978c4f, type: 3}
+      propertyPath: m_IsActive
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
