+++ Assets/Prefabs/GUI/BlackScreenFadeOutAnim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: df6a75efd0de5a54eb2686d7b6174945
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
