--- Assets/Scenes/BlackScreen.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8cbf1a07f8a2a3f4c854798e0a327c13
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
